fix dtype

+++ tests/python/pytorch/graphbolt/gb_test_utils.py
@@ -125,7 +125,7 @@ def random_homo_graphbolt_graph(
         np.arange(each_set_size),
         np.arange(each_set_size, 2 * each_set_size),
     )
-    train_data = np.vstack(train_pairs).T
+    train_data = np.vstack(train_pairs).T.astype(np.int64)
     train_path = os.path.join("set", "train.npy")
     np.save(os.path.join(test_dir, train_path), train_data)
 
@@ -133,7 +133,7 @@ def random_homo_graphbolt_graph(
         np.arange(each_set_size, 2 * each_set_size),
         np.arange(2 * each_set_size, 3 * each_set_size),
     )
-    validation_data = np.vstack(validation_pairs).T
+    validation_data = np.vstack(validation_pairs).T.astype(np.int64)
     validation_path = os.path.join("set", "validation.npy")
     np.save(os.path.join(test_dir, validation_path), validation_data)
 
@@ -141,7 +141,7 @@ def random_homo_graphbolt_graph(
         np.arange(2 * each_set_size, 3 * each_set_size),
         np.arange(3 * each_set_size, 4 * each_set_size),
     )
-    test_data = np.vstack(test_pairs).T
+    test_data = np.vstack(test_pairs).T.astype(np.int64)
     test_path = os.path.join("set", "test.npy")
     np.save(os.path.join(test_dir, test_path), test_data)
